+++ ccb.c
@@ -50,7 +50,7 @@ void GetDataFileName(Widget w, XtPointer client, XtPointer call)
 /* -------------------------------------------------------------------- */
 void ToggleProject(Widget w, XtPointer client, XtPointer call)
 {
-  if ((int)client)
+  if ((long)client)
     ProjectToBack = !ProjectToBack;
   else
     ProjectToXY = !ProjectToXY;
@@ -403,7 +403,7 @@ void ForkNetscape(Widget w, XtPointer client, XtPointer call)
 {
   if (fork() == 0)
     {
-    switch ((int)client)
+    switch ((long)client)
       {
       case 1:
         execlp("firefox", "firefox", "http://www.eol.ucar.edu/raf", NULL);
